+++ tests/testapp/test_contents.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import, unicode_literals
 
+from django.template.loader import render_to_string
 from django.test import TestCase
 
 from elephantblog.contents import BlogEntryListContent, BlogCategoryListContent
@@ -26,7 +27,7 @@ class ContentsTestCase(TestCase):
         content = BlogEntryListContent()
 
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertIn(
             'h2 class="entry-title"><a href="/multilang/2012/10/12/eintrag-1/', html,
         )
@@ -36,21 +37,21 @@ class ContentsTestCase(TestCase):
 
         content.featured_only = True
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
 
         content.category = category
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 0,
         )
 
         content.featured_only = False
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
@@ -58,40 +59,40 @@ class ContentsTestCase(TestCase):
         content = BlogEntryListContent()
         content.paginate_by = 1
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
 
         Request.GET["page"] = 2
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
 
         Request.GET["page"] = 3
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
 
         Request.GET["page"] = "abc"
         content.process(Request)
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<h2"), 1,
         )
 
         content = BlogCategoryListContent()
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<li>"), 1,
         )
 
         content.show_empty_categories = True
-        html = content.render()
+        html = render_to_string(*content.render())
         self.assertEqual(
             html.count("<li>"), 2,
         )